Move fixed_time_eq method to util module to make it public
