+++ src/rust-crypto/cryptoutil.rs
@@ -105,84 +105,6 @@ pub fn read_u32_be(input: &[u8]) -> u32 {
 }
 
 
-#[cfg(target_arch = "x86")]
-#[cfg(target_arch = "x86_64")]
-#[inline(never)]
-#[allow(dead_assignment)]
-unsafe fn fixed_time_eq_asm(mut lhsp: *const u8, mut rhsp: *const u8, mut count: uint) -> bool {
-    let mut result: u8 = 0;
-
-    asm!(
-        "
-            fixed_time_eq_loop:
-
-            mov ($1), %cl
-            xor ($2), %cl
-            or %cl, $0
-
-            inc $1
-            inc $2
-            dec $3
-            jnz fixed_time_eq_loop
-        "
-        : "+r" (result), "+r" (lhsp), "+r" (rhsp), "+r" (count) // all input and output
-        : // input
-        : "cl", "cc" // clobbers
-        : "volatile" // flags
-    );
-
-    return result == 0;
-}
-
-#[cfg(target_arch = "arm")]
-#[inline(never)]
-#[allow(dead_assignment)]
-unsafe fn fixed_time_eq_asm(mut lhsp: *const u8, mut rhsp: *const u8, mut count: uint) -> bool {
-    let mut result: u8 = 0;
-
-    asm!(
-        "
-            fixed_time_eq_loop:
-
-            ldrb r4, [$1]
-            ldrb r5, [$2]
-            eor r4, r4, r5
-            orr $0, $0, r4
-
-            add $1, $1, #1
-            add $2, $2, #1
-            subs $3, $3, #1
-            bne fixed_time_eq_loop
-        "
-        : "+r" (result), "+r" (lhsp), "+r" (rhsp), "+r" (count) // all input and output
-        : // input
-        : "r4", "r5", "cc" // clobbers
-        : "volatile" // flags
-    );
-
-    return result == 0;
-}
-
-/// Compare two vectors using a fixed number of operations. If the two vectors are not of equal
-/// length, the function returns false immediately.
-pub fn fixed_time_eq(lhs: &[u8], rhs: &[u8]) -> bool {
-    if lhs.len() != rhs.len() {
-        return false;
-    }
-    if lhs.len() == 0 {
-        return true;
-    }
-
-    let count = lhs.len();
-
-    unsafe {
-        let lhsp = lhs.unsafe_ref(0);
-        let rhsp = rhs.unsafe_ref(0);
-        return fixed_time_eq_asm(lhsp, rhsp, count);
-    }
-}
-
-
 /// symm_enc_or_dec() implements the necessary functionality to turn a SynchronousStreamCipher into
 /// an Encryptor or Decryptor
 pub fn symm_enc_or_dec<S: SynchronousStreamCipher, R: ReadBuffer, W: WriteBuffer>(
@@ -460,7 +382,7 @@ pub mod test {
     use std::rand::IsaacRng;
     use std::rand::distributions::{IndependentSample, Range};
 
-    use cryptoutil::{add_bytes_to_bits, add_bytes_to_bits_tuple, fixed_time_eq};
+    use cryptoutil::{add_bytes_to_bits, add_bytes_to_bits_tuple};
     use digest::Digest;
 
     /// Feed 1,000,000 'a's into the digest with varying input sizes and check that the result is
@@ -532,25 +454,5 @@ pub mod test {
     fn test_add_bytes_to_bits_tuple_overflow2() {
         let value: u64 = Bounded::max_value();
         add_bytes_to_bits_tuple::<u64>((value - 1, 0), 0x8000000000000000);
-    }
-
-    #[test]
-    pub fn test_fixed_time_eq() {
-        let a = [0, 1, 2];
-        let b = [0, 1, 2];
-        let c = [0, 1, 9];
-        let d = [9, 1, 2];
-        let e = [2, 1, 0];
-        let f = [2, 2, 2];
-        let g = [0, 0, 0];
-
-        assert!(fixed_time_eq(a, a));
-        assert!(fixed_time_eq(a, b));
-
-        assert!(!fixed_time_eq(a, c));
-        assert!(!fixed_time_eq(a, d));
-        assert!(!fixed_time_eq(a, e));
-        assert!(!fixed_time_eq(a, f));
-        assert!(!fixed_time_eq(a, g));
     }
 }
