+++ src/rust-crypto/mac.rs
@@ -8,7 +8,7 @@
  * The mac module defines the Message Authentication Code (Mac) trait.
  */
 
-use cryptoutil::fixed_time_eq;
+use util::fixed_time_eq;
 
 /**
  * The Mac trait defines methods for a Message Authentication function.
