+++ src/rust-crypto/pbkdf2.rs
@@ -16,10 +16,11 @@ use std::slice::MutableCloneableVector;
 use serialize::base64;
 use serialize::base64::{FromBase64, ToBase64};
 
-use cryptoutil::{read_u32_be, write_u32_be, fixed_time_eq};
+use cryptoutil::{read_u32_be, write_u32_be};
 use hmac::Hmac;
 use mac::Mac;
 use sha2::Sha256;
+use util::fixed_time_eq;
 
 // Calculate a block of the output of size equal to the output_bytes of the underlying Mac function
 // mac - The Mac function to use
